requirements.txt
requests
pyyaml
PyFunceble